+++ src/e2e/rpc/utxo/utxo.e2e.utils.ts
@@ -1,6 +1,6 @@
 import { Network } from '../../../dto'
-import { BaseUtxo, TatumSDK } from '../../../service'
-import { RpcE2eUtils } from '../rpc.e2e.utils'
+import { BaseUtxo, TatumSDK, Dogecoin } from '../../../service'
+import { e2eUtil } from '../../e2e.util'
 
 export enum UtxoNetworkType {
   MAIN = 'main',
@@ -15,8 +15,8 @@ interface TatumBtcUtils {
 }
 
 export const UtxoE2eUtils = {
-  initTatum: async <T extends BaseUtxo>(params: TatumBtcUtils) =>
-    TatumSDK.init<T>(RpcE2eUtils.initConfig(params.network, params.apiKey)),
+  initTatum: async <T extends BaseUtxo | Dogecoin>(params: TatumBtcUtils) =>
+    TatumSDK.init<T>(e2eUtil.initConfig(params.network, params.apiKey)),
   e2e: (params: TatumBtcUtils) => {
     const { type } = params
     it('chain info', async () => {